[backend] refactor: UserService를 interface와 구현체로 구조 분리
